smarteye_backend/requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
djoser==2.2.0
drf-spectacular==0.26.5
whitenoise==6.6.0

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0

# Database
psycopg2-binary==2.9.7

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# Async Task Processing
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Processing & Document Processing
Pillow==10.0.1
PyMuPDF==1.23.8
python-magic==0.4.27
reportlab==4.0.4

# AI/ML Libraries (from existing notebook)
torch==2.2.0
torchvision==0.17.0
ultralytics==8.0.200
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.3
matplotlib==3.7.2

# OCR
pytesseract==0.3.10

# API Integrations
openai==1.3.5
requests==2.31.0

# Monitoring & Logging
loguru==0.7.2
rich==13.6.0
psutil==5.9.6

# Development Tools
python-decouple==3.8
tqdm==4.66.1
huggingface-hub==0.17.3

# Web Server (Production)
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest-django==4.7.0
factory-boy==3.3.0

# Documentation
drf-spectacular==0.26.5

# File Storage (Optional - for cloud storage)
boto3==1.29.7
django-storages==1.14.2

# WebSocket support for real-time updates
channels==4.0.0
channels-redis==4.1.0

# Data validation
marshmallow==3.20.1
django-filter==23.3
